feature: ads

+++ app/src/main/java/app/coinverse/user/ProfileFragment.kt
@@ -9,7 +9,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProviders
 import androidx.navigation.findNavController
-import app.coinverse.Enums
+import app.coinverse.Enums.FeedType.DISMISSED
 import app.coinverse.R
 import app.coinverse.databinding.FragmentProfileBinding
 import app.coinverse.firebase.FirestoreCollections.usersCollection
@@ -17,7 +17,7 @@ import app.coinverse.home.HomeViewModel
 import app.coinverse.utils.DISMISS_COLLECTION
 import app.coinverse.utils.PROFILE_VIEW
 import app.coinverse.utils.SIGN_OUT_ON_BACK_PRESS_DELAY_IN_MILLIS
-import com.firebase.ui.auth.AuthUI
+import com.firebase.ui.auth.AuthUI.getInstance
 import com.google.android.material.snackbar.Snackbar
 import com.google.firebase.FirebaseApp
 import com.google.firebase.analytics.FirebaseAnalytics
@@ -27,7 +27,7 @@ import com.google.firebase.firestore.CollectionReference
 import kotlinx.android.synthetic.main.fragment_profile.*
 import kotlinx.android.synthetic.main.toolbar.*
 
-private var LOG_TAG = ProfileFragment::class.java.simpleName
+private val LOG_TAG = ProfileFragment::class.java.simpleName
 
 class ProfileFragment : Fragment() {
     private lateinit var analytics: FirebaseAnalytics
@@ -64,18 +64,17 @@ class ProfileFragment : Fragment() {
     }
 
     fun setClickListeners() {
-
         dismissedContent.setOnClickListener { view: View ->
-            val action =
-                    ProfileFragmentDirections.actionProfileFragmentToDismissedContentFragment()
-            action.feedType = Enums.FeedType.DISMISSED.name
-            view.findNavController().navigate(R.id.action_profileFragment_to_dismissedContentFragment, action.arguments)
+            view.findNavController().navigate(R.id.action_profileFragment_to_dismissedContentFragment,
+                    ProfileFragmentDirections.actionProfileFragmentToDismissedContentFragment().apply {
+                        feedType = DISMISSED.name
+                    }.arguments)
         }
 
         signOut.setOnClickListener { view: View ->
             var message: Int
             if (FirebaseAuth.getInstance().currentUser != null) {
-                AuthUI.getInstance().signOut(context!!).addOnCompleteListener {
+                getInstance().signOut(context!!).addOnCompleteListener {
                     if (it.isSuccessful) {
                         homeViewModel.user.value = null
                         message = R.string.signed_out
@@ -94,7 +93,7 @@ class ProfileFragment : Fragment() {
         delete.setOnClickListener { view: View ->
             var message: Int
             if (FirebaseAuth.getInstance().currentUser != null) {
-                AuthUI.getInstance().delete(context!!).addOnCompleteListener {
+                getInstance().delete(context!!).addOnCompleteListener {
                     if (it.isSuccessful) {
                         //TODO: Refactor to handle on server.
                         deleteCollection(usersCollection
@@ -131,19 +130,15 @@ class ProfileFragment : Fragment() {
         try {
             // Retrieve a small batch of documents to avoid out-of-memory errors/
             var deleted = 0
-            collection
-                    .limit(batchSize.toLong())
-                    .get()
-                    .addOnCompleteListener {
-                        for (document in it.result!!.documents) {
-                            document.reference.delete()
-                            ++deleted
-                        }
-                        if (deleted >= batchSize) {
-                            // retrieve and delete another batch
-                            deleteCollection(collection, batchSize)
-                        }
-                    }
+            collection.limit(batchSize.toLong()).get().addOnCompleteListener {
+                it.result!!.documents.all { document ->
+                    document.reference.delete()
+                    ++deleted
+                    true
+                }
+                // retrieve and delete another batch
+                if (deleted >= batchSize) deleteCollection(collection, batchSize)
+            }
         } catch (e: Exception) {
             System.err.println("Error deleting collection : " + e.message)
         }
